uploading authentication for all the services of TOURIST

+++ Unreal_Journey_Engine/Unreal_Journey_Engine/Controllers/FeedbackController.cs
@@ -7,6 +7,7 @@ using System.Net;
 using System.Net.Http;
 using System.Web.Http;
 using System.Web.Http.Cors;
+using Unreal_Journey_Engine.AuthFilters;
 
 namespace Unreal_Journey_Engine.Controllers
 {
@@ -19,22 +20,38 @@ namespace Unreal_Journey_Engine.Controllers
         #region Get All Feedbacks
         [HttpGet]
         [Route("all")]
+        [Logged]
         public HttpResponseMessage Get_All_Feedbacks()
         {
 
-            var data = FeedbackService.Get();
-            if (data.Count > 0)
+            var authorizationHeader = Request.Headers.Authorization?.ToString();
+            var current_user_Type = User_Info_Provider.Get_User_Role(authorizationHeader);
+            if (current_user_Type == "Admin")
             {
-                return Request.CreateResponse(HttpStatusCode.OK, data);
+                var data = FeedbackService.Get();
+                if (data.Count > 0)
+                {
+                    return Request.CreateResponse(HttpStatusCode.OK, data);
+                }
+                else
+                {
+                    var responseMessage = new
+                    {
+                        Message = "No data available"
+                    };
+                    return Request.CreateResponse(HttpStatusCode.NotFound, responseMessage);
+                }
             }
             else
             {
                 var responseMessage = new
                 {
-                    Message = "No data available"
+                    Message = "You are not allowrd to access all the feedbacks"
                 };
-                return Request.CreateResponse(HttpStatusCode.NotFound, responseMessage);
+                return Request.CreateResponse(HttpStatusCode.Forbidden, responseMessage);
             }
+
+            
 
         }
         #endregion Get All Feedbacks
@@ -42,6 +59,7 @@ namespace Unreal_Journey_Engine.Controllers
         #region Get Single Feedback
         [HttpGet]
         [Route("{id}")]
+        [Logged]
         public HttpResponseMessage Get(int id)
         {
             try
@@ -70,40 +88,55 @@ namespace Unreal_Journey_Engine.Controllers
         #region Post / Create
         [HttpPost]
         [Route("create")]
+        [Logged]
         public HttpResponseMessage Create_Feedback(FeedbackDTO dto)
         {
             try
             {
 
-                if (dto != null)
+                var authorizationHeader = Request.Headers.Authorization?.ToString();
+                var current_user_Type = User_Info_Provider.Get_User_Role(authorizationHeader);
+                if (current_user_Type == "Tourist" || current_user_Type == "Tour_Guide")
                 {
-                    var decision = FeedbackService.Create(dto);
-                    if (decision)
+                    if (dto != null)
                     {
-                        var responseMessage = new
+                        var decision = FeedbackService.Create(dto);
+                        if (decision)
                         {
-                            Message = "Feedback Posted"
-                        };
-                        return Request.CreateResponse(HttpStatusCode.OK, responseMessage);
+                            var responseMessage = new
+                            {
+                                Message = "Feedback Posted"
+                            };
+                            return Request.CreateResponse(HttpStatusCode.OK, responseMessage);
+                        }
+                        else
+                        {
+                            var responseMessage = new
+                            {
+                                Message = "Failed to Create Feedback"
+                            };
+                            return Request.CreateResponse(HttpStatusCode.NotAcceptable, responseMessage);
+                        }
+
                     }
                     else
                     {
                         var responseMessage = new
                         {
-                            Message = "Failed to Create Feedback"
+                            Message = "Provide Feedback Data to Create Feedback"
                         };
-                        return Request.CreateResponse(HttpStatusCode.NotAcceptable, responseMessage);
+                        return Request.CreateResponse(HttpStatusCode.PreconditionFailed, responseMessage);
                     }
-
                 }
                 else
                 {
                     var responseMessage = new
                     {
-                        Message = "Provide Feedback Data to Create Feedback"
+                        Message = "You are not allowed to send any feedbacks. Only Tourists and Tour Guides can"
                     };
-                    return Request.CreateResponse(HttpStatusCode.PreconditionFailed, responseMessage);
+                    return Request.CreateResponse(HttpStatusCode.Forbidden, responseMessage);
                 }
+                
             }
             catch (Exception ex)
             {
@@ -115,6 +148,7 @@ namespace Unreal_Journey_Engine.Controllers
         #region Update
         [HttpPut]
         [Route("update")]
+        [Logged]
         public HttpResponseMessage Update_Feedback_Info(FeedbackDTO dto)
         {
             try
@@ -122,23 +156,37 @@ namespace Unreal_Journey_Engine.Controllers
 
                 if (dto != null)
                 {
-                    var decision = FeedbackService.Update(dto);
-                    if (decision)
+                    var authorizationHeader = Request.Headers.Authorization?.ToString();
+                    var current_user_Type = User_Info_Provider.Get_User_Role(authorizationHeader);
+                    if (current_user_Type == "Admin")
                     {
-                        var responseMessage = new
+                        var decision = FeedbackService.Update(dto);
+                        if (decision)
                         {
-                            Message = "Feedback Updated"
-                        };
-                        return Request.CreateResponse(HttpStatusCode.OK, responseMessage);
+                            var responseMessage = new
+                            {
+                                Message = "Feedback Updated"
+                            };
+                            return Request.CreateResponse(HttpStatusCode.OK, responseMessage);
+                        }
+                        else
+                        {
+                            var responseMessage = new
+                            {
+                                Message = "Failed to Update Feedback"
+                            };
+                            return Request.CreateResponse(HttpStatusCode.NotAcceptable, responseMessage);
+                        }
                     }
                     else
                     {
                         var responseMessage = new
                         {
-                            Message = "Failed to Update Feedback"
+                            Message = "You are not allowed to Update Feedbacks. Only admin can"
                         };
-                        return Request.CreateResponse(HttpStatusCode.NotAcceptable, responseMessage);
+                        return Request.CreateResponse(HttpStatusCode.Forbidden, responseMessage);
                     }
+                    
 
                 }
                 else
@@ -160,22 +208,36 @@ namespace Unreal_Journey_Engine.Controllers
         #region Delete
         [HttpDelete]
         [Route("delete/{id}")]
+        [Logged]
         public HttpResponseMessage Delete_Feedback_Info(int id)
         {
             try
             {
-                var data = FeedbackService.Delete(id);
-                if (data)
+                var authorizationHeader = Request.Headers.Authorization?.ToString();
+                var current_user_Type = User_Info_Provider.Get_User_Role(authorizationHeader);
+                if(current_user_Type == "Admin")
                 {
-                    return Request.CreateResponse(HttpStatusCode.OK, data);
+                    var data = FeedbackService.Delete(id);
+                    if (data)
+                    {
+                        return Request.CreateResponse(HttpStatusCode.OK, data);
+                    }
+                    else
+                    {
+                        var responseMessage = new
+                        {
+                            Message = "Feedback Not Found"
+                        };
+                        return Request.CreateResponse(HttpStatusCode.NotFound, responseMessage);
+                    }
                 }
                 else
                 {
                     var responseMessage = new
                     {
-                        Message = "Feedback Not Found"
+                        Message = "You are not allowed to delete feedbacks. Only admin can."
                     };
-                    return Request.CreateResponse(HttpStatusCode.NotFound, responseMessage);
+                    return Request.CreateResponse(HttpStatusCode.Forbidden, responseMessage);
                 }
             }
             catch (Exception ex)
